requirements.txt
pyrebase4
python-dotenv